Massive refactoring Code cleanup Deckerization
