+++ pom.xml
@@ -5,14 +5,14 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.3.5.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<version>2.4.0</version>
+		<relativePath/>
 	</parent>
-	<groupId>com.cs322</groupId>
-	<artifactId>AuthService</artifactId>
+	<groupId>am.aua.cs322</groupId>
+	<artifactId>orgchart</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>AuthService</name>
-	<description>Demo project for Spring Boot</description>
+	<name>AUA Org Chart</name>
+	<description>AUA CS322 course project - Organizational Chart</description>
 
 	<properties>
 		<java.version>11</java.version>
