+++ src/main/java/am/aua/cs322/orgchart/model/User.java
@@ -1,4 +1,4 @@
-package com.cs322.AuthService.Model;
+package am.aua.cs322.orgchart.model;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
