+++ src/main/java/am/aua/cs322/orgchart/security/JwtInMemoryUserDetailsService.java
@@ -1,19 +1,23 @@
-package com.cs322.AuthService.Security;
+package am.aua.cs322.orgchart.security;
 
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.util.*;
-
-import com.cs322.AuthService.Model.User;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.*;
+import am.aua.cs322.orgchart.model.User;
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParser;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
+import java.io.BufferedReader;
+import java.io.FileNotFoundException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
 
 @Service
 public class JwtInMemoryUserDetailsService implements UserDetailsService {
@@ -22,7 +26,7 @@ public class JwtInMemoryUserDetailsService implements UserDetailsService {
 
     @PostConstruct
     private void getAllUsers() throws FileNotFoundException {
-        JsonElement root = new JsonParser().parse(new FileReader("src/main/java/com/cs322/AuthService/Security/users.json"));
+        JsonElement root = new JsonParser().parse(getReaderFromFileName("/data/users.json"));
         JsonArray object = root.getAsJsonArray();
 
         Gson gson = new Gson();
@@ -41,6 +45,17 @@ public class JwtInMemoryUserDetailsService implements UserDetailsService {
         }
 
         return user.get();
+    }
+
+    /**
+     * Returns BufferedReader for give file. Other methods fail when the application is packages as Jar
+     *
+     * @param path relative path of the file to the classpath
+     * @return BufferedReader for current file
+     */
+    private BufferedReader getReaderFromFileName(String path) {
+        InputStream in = getClass().getResourceAsStream(path);
+        return new BufferedReader(new InputStreamReader(in));
     }
 }
